Cast one of the operands of this addition operation to a "float".

+++ src/jmemorize/gui/swing/panels/CardCounterPanel.java
@@ -281,7 +281,7 @@ public class CardCounterPanel extends JPanel
 
                 frontLearnedAmount = Math.min(frontLearnedAmount, frontTargetAmount);
                 backLearnedAmount = Math.min(backLearnedAmount, backTargetAmount);
-                vals[i] = (frontLearnedAmount + backLearnedAmount) / targetAmount;
+                vals[i] = ((float)frontLearnedAmount + backLearnedAmount) / targetAmount;
             }
         }
